==> notebooks/1_Postcode_import.ipynb <==
### Notebook to Import Postcode ###
We import a postcode dataset found online containing useful information such as latitude, longitude and suburb names (aka locality)

# Setting up Spark Session

from pyspark.sql import SparkSession
spark = SparkSession.builder.appName("CSV to Parquet").getOrCreate()
df = spark.read.csv("../data/postcodes/postcodes.csv", header=True, inferSchema=True)

#Set up necessary files
from urllib.request import urlretrieve
import os

# from the current `tute_1` directory, go back two levels to the `MAST30034` directory
output_relative_dir = '../data/'

# check if it exists as it makedir will raise an error if it does exist
if not os.path.exists(output_relative_dir):
    os.makedirs(output_relative_dir)
    
if not os.path.exists(output_relative_dir + 'postcodes'):
    os.makedirs(output_relative_dir + 'postcodes')

url = "https://www.matthewproctor.com/Content/postcodes/australian_postcodes.csv"
postcodes_output_dir = output_relative_dir + 'postcodes'


# generate output location and filename
output_dir = "postcodes.csv"
output_dir = f"{postcodes_output_dir}/postcodes.csv"

import urllib.request

# bypass Error 403 : Forbidden Access
opener = urllib.request.URLopener()
opener.addheader('User-Agent', 'whatever')  
filename, headers = opener.retrieve(url, output_dir)

We then filtered for Victorian suburbs

vic_df = df.filter(df['state'] == 'VIC')
vic_df.show()

Unnecessary columns were also removed

columns = ['id','state','dc','type','status','sa3','sa3name','sa4','sa4name','region','Lat_precise',
            'Long_precise','SA1_CODE_2021','SA1_NAME_2021','SA2_CODE_2021','SA2_NAME_2021','SA3_CODE_2021',
            'SA3_NAME_2021','SA4_CODE_2021','SA4_NAME_2021','RA_2011','RA_2016','RA_2021','RA_2021_NAME',
            'MMM_2015','MMM_2019','ced','altitude','chargezone','phn_code','phn_name','lgaregion','lgacode',
            'electorate','electoraterating','sed_code','sed_name']

vic_df = vic_df.drop(*columns)
vic_df.show()

We ensured that there were no federal postcodes in the dataframe

vic_df = vic_df.filter(df['postcode'] < 4000)

Finally, we saved the resulting dataframe in a parquet file

if not os.path.exists("../data/postcodes/postcodes.parquet"):
    vic_df.write.parquet("../data/postcodes/postcodes.parquet")
