==> notebooks/5_Liveability_index.ipynb <==
## Liveability Model
How liveable is a suburb, really?

We used our liveability metrics of: 
- proximity to the city
- groceries access
- healthcare access
- education access
- affordability 

All these liveability metrics are numerical data on a scale of 1 - 10 for quanitifable anaylsis and interpretability

# Import necessary libraries 
import pandas as pd
import numpy as np
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from pyspark.ml.feature import Bucketizer
from pyspark.sql.functions import col, desc


# Create a spark session (which will run spark jobs)
spark = (
    SparkSession.builder.appName("Liveability")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config('spark.driver.memory', '4g')
    .config('spark.executor.memory', '2g')
    .getOrCreate()
)

The dataframe containing all the liveability metrics is imported

sdf = spark.read.parquet('../data/curated/suburb_level_data.parquet', header=True, inferSchema=True)
sdf.show()

### On a scale of 1 - 10... 
All the data was placed on a scale of 1-10. Unfortunately, we could not find any data on what number of schools, groceries, healthcare or education per capita was ideal. We could, however, compare how well equipped one suburb was compared to other suburbs in Victoria. Hence, for all liveability variables, we bucketised each of the per capita data by finding each 10th percentile and assigning a respective score out of 10 for it. For instance if the groceries per capita of a particular suburb is in the top 10th percentile, the suburb would receive a 10/10 score. 

######## Scale of 1 - 10 for Schools ########

# Calculate quantile cut points
quantiles = sdf.approxQuantile("school_per_capita", [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9], 0.0)

quantiles = sorted(set(quantiles))

# Add min and max to make splits
splits = [-float('inf')] + quantiles + [float('inf')]

# Create the Bucketizer
bucketizer = Bucketizer(
    splits=splits,
    inputCol="school_per_capita",
    outputCol="school_per_capita_score"
)


######## Scale of 1 - 10 for Groceries ########
quantiles2 = sdf.approxQuantile("groceries_per_capita", [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9], 0.0)
# Add min and max to make splits
splits2 = [-float('inf')] + quantiles2 + [float('inf')]

# Create the Bucketizer
bucketizer2 = Bucketizer(
    splits=splits,
    inputCol="groceries_per_capita",
    outputCol="groceries_per_capita_score"
)


######## Scale of 1 - 10 for Healthcare ########
quantiles3 = sdf.approxQuantile("healthcare_per_capita", [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9], 0.0)
# Add min and max to make splits
splits3 = [-float('inf')] + quantiles3 + [float('inf')]

# Create the Bucketizer
bucketizer3 = Bucketizer(
    splits=splits,
    inputCol="healthcare_per_capita",
    outputCol="healthcare_per_capita_score"
)

# Transform the DataFrame to include all 3 caregorical metric
sdf_buckets = bucketizer.transform(sdf)
sdf_buckets2 = bucketizer2.transform(sdf_buckets)
sdf_buckets3 = bucketizer3.transform(sdf_buckets2)

sdf_buckets3.show()

######## Scale of 1 - 10 for Proximity to the City ########
sdf_buckets4 = sdf_buckets3 .withColumn(
    "distance_score",
    F.when((F.col("distance_to_melbourne_km") <= 5), 10)
    .when((F.col("distance_to_melbourne_km") > 5) & (F.col("distance_to_melbourne_km") <= 10),9 )
    .when((F.col("distance_to_melbourne_km") > 10) & (F.col("distance_to_melbourne_km") <= 15), 8)
    .when((F.col("distance_to_melbourne_km") > 15) & (F.col("distance_to_melbourne_km") <= 20), 7)
    .when((F.col("distance_to_melbourne_km") > 20) & (F.col("distance_to_melbourne_km") <= 25), 6)
    .when((F.col("distance_to_melbourne_km") > 25) & (F.col("distance_to_melbourne_km") <= 30), 5)
    .when((F.col("distance_to_melbourne_km") > 30) & (F.col("distance_to_melbourne_km") <= 35), 4)
    .when((F.col("distance_to_melbourne_km") > 35) & (F.col("distance_to_melbourne_km") <= 40), 3)
    .when((F.col("distance_to_melbourne_km") > 40) & (F.col("distance_to_melbourne_km") <= 45), 2)
    .when((F.col("distance_to_melbourne_km") > 50) , 1),
    )
sdf_buckets4

For afforability, the RAI (rent affordability index) was used to classify the affordability of each particular type of property across all suburbs. This metric was based on another paper online: *https://sgsep.com.au/projects/rental-affordability-index*

# Across all types of properties
sdf_buckets5 = sdf_buckets4.withColumn(
    "all_RAI_score",
    F.when((F.col("all_RAI") <= 50), 0)
    .when((F.col("all_RAI") > 50) & (F.col("all_RAI") <= 75), 1) 
    .when((F.col("all_RAI") > 75) & (F.col("all_RAI") <= 100), 2)
    .when((F.col("all_RAI") > 100) & (F.col("all_RAI") <= 115), 3)
    .when((F.col("all_RAI") > 115) & (F.col("all_RAI") <= 130), 4)
    .when((F.col("all_RAI") > 130) & (F.col("all_RAI") <= 145), 5)
    .when((F.col("all_RAI") > 145) & (F.col("all_RAI") <= 160), 6)
    .when((F.col("all_RAI") > 160) & (F.col("all_RAI") <= 175), 7)
    .when((F.col("all_RAI") > 150) & (F.col("all_RAI") <= 175), 8)
    .when((F.col("all_RAI") > 175) & (F.col("all_RAI") <= 200), 9)
    .when((F.col("all_RAI") > 200), 10),
    )

# Across 1 - 2 bedroom properties
sdf_buckets6 = sdf_buckets5.withColumn(
    "1-2_Bed_RAI_score",
    F.when((F.col("1-2_Bed_RAI") <= 50), 0) 
    .when((F.col("1-2_Bed_RAI") > 50) & (F.col("1-2_Bed_RAI") <= 75), 1) 
    .when((F.col("1-2_Bed_RAI") > 75) & (F.col("1-2_Bed_RAI") <= 100), 2) 
    .when((F.col("1-2_Bed_RAI") > 100) & (F.col("1-2_Bed_RAI") <= 115), 3) 
    .when((F.col("1-2_Bed_RAI") > 115) & (F.col("1-2_Bed_RAI") <= 130), 4) 
    .when((F.col("1-2_Bed_RAI") > 130) & (F.col("1-2_Bed_RAI") <= 145), 5) 
    .when((F.col("1-2_Bed_RAI") > 145) & (F.col("1-2_Bed_RAI") <= 160), 6) 
    .when((F.col("1-2_Bed_RAI") > 160) & (F.col("1-2_Bed_RAI") <= 175), 7) 
    .when((F.col("1-2_Bed_RAI") > 150) & (F.col("1-2_Bed_RAI") <= 175), 8) 
    .when((F.col("1-2_Bed_RAI") > 175) & (F.col("1-2_Bed_RAI") <= 200), 9) 
    .when((F.col("1-2_Bed_RAI") > 200),10),
    )

# Across 3+ bedroom properties
sdf_buckets7 = sdf_buckets6.withColumn(
    "3+_Bed_RAI_score",
    F.when((F.col("3+_Bed_RAI") <= 50), 0) 
    .when((F.col("3+_Bed_RAI") > 50) & (F.col("3+_Bed_RAI") <= 75), 1) 
    .when((F.col("3+_Bed_RAI") > 75) & (F.col("3+_Bed_RAI") <= 100), 2) 
    .when((F.col("3+_Bed_RAI") > 100) & (F.col("3+_Bed_RAI") <= 115), 3) 
    .when((F.col("3+_Bed_RAI") > 115) & (F.col("3+_Bed_RAI") <= 130), 4) 
    .when((F.col("3+_Bed_RAI") > 130) & (F.col("3+_Bed_RAI") <= 145), 5) 
    .when((F.col("3+_Bed_RAI") > 145) & (F.col("3+_Bed_RAI") <= 160), 6) 
    .when((F.col("3+_Bed_RAI") > 160) & (F.col("3+_Bed_RAI") <= 175), 7) 
    .when((F.col("3+_Bed_RAI") > 150) & (F.col("3+_Bed_RAI") <= 175), 8) 
    .when((F.col("3+_Bed_RAI") > 175) & (F.col("3+_Bed_RAI") <= 200), 9) 
    .when((F.col("3+_Bed_RAI") > 200), 10)
    )
sdf_buckets7

We then split the dataset according to the property types. Duplicates and unnecessary columns were also removed.


sdf1 = sdf_buckets7.filter(sdf['bed_column'] == '1-2_bedders')
sdf1 = sdf1.distinct()
sdf2 = sdf_buckets7.filter(sdf['bed_column'] == '3+_bedders')
sdf2 = sdf2.distinct()
sdf_1a = sdf1.drop(sdf1['3+_Bed_RAI'], sdf1['bed_column'], sdf1['groceries_per_capita'], sdf1['school_per_capita'], sdf1['healthcare_per_capita'],
        sdf1['total population - 2021'], sdf1['Latitude'], sdf1['Longitude'], sdf1['distance_to_melbourne_km']) 
sdf_2a = sdf2.drop(sdf2['1-2_Bed_RAI'], sdf2['bed_column'], sdf2['groceries_per_capita'], sdf2['school_per_capita'], sdf2['healthcare_per_capita'],
        sdf2['total population - 2021'], sdf2['Latitude'], sdf2['Longitude'], sdf2['distance_to_melbourne_km'])
sdf_1a

### Liveability without weightages 
Firstly we designed a liveability index that weighted all variables equally. As the total weightage adds up to 10 , we assigned  a weightage of 2 for each variable

df_all = sdf_1a.select(
    col("postcode"),
    col("suburb"),
    col("all_RAI_score"),
    col("school_per_capita_score"),
    col("groceries_per_capita_score"),
    col("healthcare_per_capita_score"),
    col("distance_score"),
    (2*col("all_RAI_score") + 2*col("school_per_capita_score")+  2*col("groceries_per_capita_score")
    +  2*col("healthcare_per_capita_score")+  2*col("distance_score")).alias("liveablity_score_all")
)
df_all = df_all.drop(df_all['all_RAI_score'],df_all['school_per_capita_score'], df_all['groceries_per_capita_score'],
        df_all['healthcare_per_capita_score'],df_all['distance_score'])



### Liveability: 1 - 2 Bedrooms 
We designed a liveability index that weighted variables for 1 - 2 Bedroom properties according to what we think a 20-something would prioritise.
- 40%: affordability
- 20%: groceries
- 20%: proximity to the city
- 10%: education
- 10%: healthcare

df_1_2B = sdf_1a.select(
    col("postcode"),
    col("suburb"),
    col("1-2_Bed_RAI_score"),
    col("school_per_capita_score"),
    col("groceries_per_capita_score"),
    col("healthcare_per_capita_score"),
    col("distance_score"),
    (4*col("1-2_Bed_RAI_score") +  1*col("school_per_capita_score")+  2*col("groceries_per_capita_score")
    +  1*col("healthcare_per_capita_score")+  2*col("distance_score")).alias("liveablity_score_1_2Bedder")
)

df_1_2B = df_1_2B.drop(df_1_2B['1-2_Bed_RAI_score'],df_1_2B['school_per_capita_score'], df_1_2B['groceries_per_capita_score'],
        df_1_2B['healthcare_per_capita_score'],df_1_2B['distance_score'])

df_1_2B

### Liveability: 3+ Bedrooms 
We designed a liveability index that weighted variables for 3+ Bedroom properties according to what we think families would prioritise.
- 40%: affordability
- 20%: groceries
- 20%: education
- 10%: proximity to the city
- 10%: healthcare


df_3B = sdf_2a.select(
    col("postcode"),
    col("suburb"),
    col("3+_Bed_RAI_score"),
    col("school_per_capita_score"),
    col("groceries_per_capita_score"),
    col("healthcare_per_capita_score"),
    col("distance_score"),
    (4*col("3+_Bed_RAI_score") +  2*col("school_per_capita_score")+ 2*col("groceries_per_capita_score")
    +  1*col("healthcare_per_capita_score")+ 1*col("distance_score")).alias("liveablity_score_3+Bedder")
)

df_3B = df_3B.drop(df_3B['3+_Bed_RAI_score'],df_3B['school_per_capita_score'], df_3B['groceries_per_capita_score'],
        df_3B['healthcare_per_capita_score'],df_3B['distance_score'])

Finally, we collated all of our findings in one dataframe

liveability_sdf = df_all.join(df_1_2B, on=['postcode', 'suburb'], how='outer')
liveability_sdf1 = liveability_sdf.join(df_3B, on=['postcode', 'suburb'], how='outer')
liveability_sdf1 = liveability_sdf1.distinct()  
liveability_sdf2 = liveability_sdf1.dropna(subset=['postcode', 'suburb'])
liveability_sdf2 = liveability_sdf1.dropDuplicates(subset=['postcode', 'suburb'])
liveability_sdf2

### Most Liveable Suburbs Across All Property Types

sorted_df_all = df_all.orderBy(desc("liveablity_score_all"))
sorted_df_all.show()

### Most Liveable Suburbs for 1 - 2 Bedrooms

sorted_df_1_2B = df_1_2B.orderBy(desc('liveablity_score_1_2Bedder'))
sorted_df_1_2B.show()

### Most Liveable Suburb for 3+ Bedrooms

sorted_df_3B = df_3B.orderBy(desc('liveablity_score_3+Bedder'))
sorted_df_3B.show()

## Affordability ~ $$$ ~ 
How affordable is a suburb, really? Although we had used the RAI to calculate the liveability of suburbs, we have decided to better visualise affordability by comparing the median rent of each suburb with the median household income of Victorians. Affordable rent is rent within 30% of the household income. The equation for this is (100 / RAI)

percentage_sdf = sdf_buckets7.select(
    col("postcode"),
    col("suburb"),
    col("all_RAI"),
    col("1-2_Bed_RAI"),
    col("3+_Bed_RAI"),
    ((100/col("all_RAI")).alias("Percentage Income : All")),
    ((100/col("1-2_Bed_RAI")).alias("Percentage Income : 1 - 2 Bedders")),
    ((100/col("3+_Bed_RAI")).alias("Percentage Income : 3 Bedders"))
)

percentage_sdf = percentage_sdf.dropDuplicates(['suburb'])
percentage_sdf


# df_3B = sdf_2a.select(
#     col("postcode"),
#     col("suburb"),
#     col("3+_Bed_RAI_score"),
#     col("school_per_capita_score"),
#     col("groceries_per_capita_score"),
#     col("healthcare_per_capita_score"),
#     col("distance_score"),
#     (4*col("3+_Bed_RAI_score") +  4*col("school_per_capita_score")+ 2*col("groceries_per_capita_score")
#     +  1*col("healthcare_per_capita_score")+ 1*col("distance_score")).alias("liveablity_score_3+Bedder")
# )

### Affordability: All Property Types

all_affordability_sdf1 = percentage_sdf.select("postcode", "suburb", "Percentage Income : All").orderBy(desc("Percentage Income : All"))
all_affordability_sdf1 = all_affordability_sdf1.dropDuplicates()
all_affordability_sdf1.orderBy(desc("Percentage Income : All"))

all_affordability_sdf1.orderBy(("Percentage Income : All"))

### Affordability : 1-2 Bedrooms

affordability_1B_sdf1 = percentage_sdf.select("postcode", "suburb", "Percentage Income : 1 - 2 Bedders").orderBy(desc("Percentage Income : 1 - 2 Bedders"))
affordability_1B_sdf1  = affordability_1B_sdf1.dropDuplicates()
affordability_1B_sdf1.orderBy(desc("Percentage Income : 1 - 2 Bedders"))

affordability_1B_sdf2 = affordability_1B_sdf1.dropna()
affordability_1B_sdf2.orderBy(("Percentage Income : 1 - 2 Bedders"))

### Affordability : 3+ Bedrooms

affordability_3B_sdf1 = percentage_sdf.select("postcode", "suburb", "Percentage Income : 3 Bedders").orderBy(desc("Percentage Income : 3 Bedders"))
affordability_3B_sdf2 = (affordability_3B_sdf1.dropDuplicates()).dropna()
affordability_3B_sdf2.orderBy(desc("Percentage Income : 3 Bedders")).show()



affordability_3B_sdf3 = (percentage_sdf.select("postcode", "suburb", "Percentage Income : 3 Bedders").orderBy(("Percentage Income : 3 Bedders")))
affordability_3B_sdf4 = (affordability_3B_sdf3.dropDuplicates()).dropna()
affordability_3B_sdf4.orderBy(("Percentage Income : 3 Bedders")).show()

## Combining our data
We decided to combine the liveability scores and affordabilit of each suburb into one single dataframe

output = liveability_sdf2.join(percentage_sdf, on = ['postcode','suburb'], how = 'inner')
output

# Unnecessary columns and duplicates were removed a
final_output = output.drop(output['all_RAI'],output['1-2_Bed_RAI'],output['3+_Bed_RAI'])
final_output = final_output.dropDuplicates(['suburb'])
final_output
