==> notebooks/6_Preprocess_suburb_level.ipynb <==
## Pre-processing on a Suburb Level
#### Liveability Metrics : Healthcare / Education / Groceries

The necessary  libraries were imported and a Spark session was initiated

# import necessary libraries
import pandas as pd
import numpy as np
from pyspark.sql import SparkSession
# Create a spark session (which will run spark jobs)
spark = (
    SparkSession.builder.appName("Liveability")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config('spark.driver.memory', '4g')
    .config('spark.executor.memory', '2g')
    .getOrCreate()
)

The dataset that contained the number of schools, groceries and healthcare was imported.


df = spark.read.parquet('../data/raw/combined_data_with_geolocation.parquet', header=True, inferSchema=True)
# Print a sample output
df.show()

#### Liveability Metrics : Affordability
The RAI dataframe was also imported to be combined with the previous dataframe

RAI_df = spark.read.parquet('../data/raw/RAI.parquet')

# Print some sample outputs from the dataframe
print((RAI_df))

The column names were renamed to follow snake_case for consistency

df = df.withColumnRenamed("School_Count", "number_of_schools")
df = df.withColumnRenamed("groceries_Count", "number_of_groceries")
df = df.withColumnRenamed("hc_Count", "number_of_healthcare_services")
df = df.withColumnRenamed("Postcode_Name", "suburbs1")
df

The two dataframes were joined on their postcodes and duplicate columns were dropped

output = df.join(RAI_df,['Postcode'],"inner")
output1 = output.drop(output['suburbs1'],output['suburbs'],output['latitude'],output['longitude'] )
output1.show()

The average rent affordability was calculated according to the porperty type as such : 
- 1 - 2 Bedrooms : (*1 bedroom flats* + *2 bedroom flats* + *3 bedroom flats*) / 3 
- 3+ Bedrooms: (*3 bedroom flats* + *3 bedroom houses* + *4 bedroom houses*) / 3

from pyspark.sql.functions import expr

# Combines all 1 - 2 Bedroom properties together 
output1 = output1.withColumn("1-2_Bed_RAI", expr("(f1B_RAI + f2B_RAI + h2B_RAI) / 3"))

# Combines all 3+ Bedroom properties together 
output1 = output1.withColumn("3+_Bed_RAI", expr("(f3B_RAI + h3B_RAI + h4B_RAI) / 3"))
output1.show()

# Previous individual RAI columns were dropped for simplicity
output2 = output1.drop( output1['f1B_RAI'], output1['f2B_RAI'],output1['f3B_RAI'] ,
                    output1['h2B_RAI'], output1['h3B_RAI'] ,output1['h4B_RAI'], output1['__index_level_0__'])
output2.show()

#### Liveability Metrics : Groceries, Hea
The liveability dataset was imported into a Spark dataframe.
However, as this dataset was combined on a property level, further pre-processing was required. 

df = spark.read.parquet('../data/landing/liveability_data.parquet')
df.show()

Unnecessary columns of the liveability dataframe was dropped

df1 = df.drop (df['cost'],df['furnished'],df['baths'],df['parking'],df['region'],df['lgaregion'],
                df['total male population - 2021'],df['total female population - 2021'], 
                df['australian citizens'], df['median rent'], df['median family weekly income'], df['median age'],
                df['total region male population - 2022'], df['total region female population - 2022'], df['Postcode_Name'])
df1.show()

All duplicate rows were removed

df2 = df1.drop_duplicates()
df2.show() 

The property type was converted to lower case for consistency and more accurate string matching

from pyspark.sql.functions import lower
df2 = df2.withColumn("property_type", lower(df["property_type"]))
df2

The number of beds for each entry was calculated and categorised into: 
- 1-2 bedrooms
- 3+ bedrooms

from pyspark.sql import functions as F

# Create the bed_column in sdf1
df3 = df2.withColumn(
    "bed_column",
    F.when((F.col("beds") == 1)  | (F.col("beds") == 2), "1-2_bedders")
    .when(F.col("beds") >= 3, "3+_bedders")
    )
df3


# Removed unecessary columns
df4 = df3.drop(df3['property_type'],df3['beds'])

### Livebility Metrics: Groceries, Healthcare per Capita
The number of groceries, healthcare per capita were calculated and added

from pyspark.sql.functions import col
from pyspark.sql import functions as F

# Calculate the Groceries per Capita
df_g = df4.select(
    col("postcode"),
    col("groceries_Count"),
    col("total population - 2021"),
    (col("groceries_Count") / col("`total population - 2021`")).alias("groceries_per_capita")
)
new_df3 = df4.join(df_g,on=['postcode','groceries_Count','total population - 2021'], how='inner')

# Calculate the Healthcare per Capita
df_h = df4.select(
    col("postcode"),
    col("hc_Count"),
    col("total population - 2021"),
    (col("hc_Count") / col("`total population - 2021`")).alias("healthcare_per_capita")
)
df_temp = df4.join(df_h,on=['postcode','hc_Count','total population - 2021'], how='inner')
df5 = df_temp.join(df_g,on=['postcode','groceries_Count','total population - 2021'], how='inner') 

Duplicate entries and unnecessary columns are removed

# Remove duplicates from the per capita dataframe 
df_no_duplicates = df5.distinct()
df_no_duplicates

# Remove unnecessary columns from the dataframe with categorised RAi according to property types  
output3 = output2.drop(output2['number_of_schools'],output2['number_of_groceries'],output2['number_of_healthcare_services'])

All the liveability metrics data was combined into one single dataframe and duplicate columns were dropped. 

# The liveability metrics dataframe containg groceries and healthcare per capita  
sdf = df_no_duplicates.join(output3, on=['postcode'], how='inner')
sdf1 = sdf.drop(sdf['hc_Count'],sdf['School_Count'],sdf['groceries_Count'])
sdf1 = sdf1.dropDuplicates(['postcode'])
sdf1

The resulting dataframe was output and saved

# the final dataframe was output into a separate file
sdf1.write.mode("overwrite").parquet('../data/landing/suburb_level_data.parquet')
